+++ screens/Stacks/Setting.jsx
@@ -3,6 +3,7 @@ import Security from '../Setting/Security';
 import Mood from '../Setting/Mood';
 import Rappel from '../Setting/Rappel';
 import Home from '../Setting/Home';
+import Review from '../Setting/Review';
 
 const stack = createNativeStackNavigator();
 
@@ -17,6 +18,7 @@ const SettingStack = () => {
       <stack.Screen name="Rappel" component={Rappel} />
       <stack.Screen name="Security" component={Security} />
       <stack.Screen name="Mood" component={Mood} />
+      <stack.Screen name="Review" component={Review} />
     </stack.Navigator>
   );
 };
